john-preacher/src/requirements.txt
torch
pandas
transformers
gitpython
annoy
